fix: use base view model, increase category counter when task added, refactor state message response
