+++ app/src/main/java/com/london/tudee/di/ViewModelModule.kt
@@ -1,18 +1,17 @@
 package com.london.tudee.di
 
-import com.london.tudee.presentation.screens.home.HomeViewModel
-import com.london.tudee.presentation.screens.task.confirm_delete_task.ConfirmDeleteTaskViewModel
-import com.london.tudee.presentation.screens.task.taskdetails.TaskDetailsBottomSheetViewModel
-import com.london.tudee.presentation.screens.task.view_tasks.EditTaskViewModel
-import com.london.tudee.presentation.screens.tasks.TasksScreenViewModel
 import com.london.tudee.presentation.screens.categories.CategoriesViewModel
 import com.london.tudee.presentation.screens.categories.CreateCategoryViewModel
 import com.london.tudee.presentation.screens.categories.EditCategoryScreenViewModel
 import com.london.tudee.presentation.screens.categories.crud.CreateCategoryScreenViewModel
 import com.london.tudee.presentation.screens.categories.crud.DeleteCategoryScreenViewModel
+import com.london.tudee.presentation.screens.home.HomeViewModel
 import com.london.tudee.presentation.screens.onboarding.OnBoardingViewModel
 import com.london.tudee.presentation.screens.onboarding.OnboardingPreferences
-import com.london.tudee.presentation.screens.task.add_edit_task_bottom_sheet.AddOrEditTaskViewModel
+import com.london.tudee.presentation.screens.task.confirm_delete_task.ConfirmDeleteTaskViewModel
+import com.london.tudee.presentation.screens.task.taskdetails.TaskDetailsBottomSheetViewModel
+import com.london.tudee.presentation.screens.task.view_tasks.EditTaskViewModel
+import com.london.tudee.presentation.screens.tasks.TasksScreenViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
@@ -28,6 +27,5 @@ val viewModelModule = module {
     viewModelOf(::CreateCategoryScreenViewModel)
     viewModelOf(::DeleteCategoryScreenViewModel)
     viewModelOf(::OnBoardingViewModel)
-    viewModelOf(::AddOrEditTaskViewModel)
     single { OnboardingPreferences(get()) }
 }
