+++ app/src/main/java/com/london/tudee/presentation/base/BaseCreateTaskInteractions.kt
@@ -17,4 +17,5 @@ interface BaseCreateTaskInteractions {
     fun saveTask()
     fun validateForm()
     fun clearMessages()
+    fun initializeForEdit(taskId: Int)
 }
