+++ app/src/main/java/com/london/tudee/presentation/screens/task/add_edit_task_bottom_sheet/AddOrEditTaskDetails.kt
@@ -1,5 +1,6 @@
 package com.london.tudee.presentation.screens.task.add_edit_task_bottom_sheet
 
+import android.annotation.SuppressLint
 import androidx.annotation.StringRes
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -38,6 +39,7 @@ import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
 
+@SuppressLint("ConfigurationScreenWidthHeight")
 @Composable
 fun AddOrEditTaskDetails(
     modifier: Modifier = Modifier,
@@ -220,7 +222,6 @@ private fun CategoriesGrid(
     selectedCategory: Category?,
     onCategorySelected: (Category) -> Unit
 ) {
-    // Memoize the chunked categories to prevent recalculation
     val chunkedCategories = remember(categories) {
         categories.chunked(3)
     }
@@ -273,106 +274,95 @@ private fun PreviewCategorySection() {
                 .padding(16.dp)
         ) {
             var selectedCategory by remember { mutableStateOf<Category?>(null) }
-            val sampleCategories = rememberSampleDomainCategories()
 
             CategorySection(
-                categories = sampleCategories,
+                categories = listOf(
+                    Category(
+                        id = 1,
+                        title = "Education",
+                        // arName = "التعليم",
+                        iconRes = "",
+                        isDefault = true,
+                        taskCount = 0,
+                        // tint = primaryColor
+                    ),
+                    Category(
+                        id = 2,
+                        title = "Shopping",
+                        // arName = "التسوق",
+                        iconRes = "",
+                        isDefault = true,
+                        taskCount = 0,
+                        // tint = secondaryColor
+                    ),
+                    Category(
+                        id = 3,
+                        title = "Medical",
+                        // arName = "طبي",
+                        iconRes = "",
+                        isDefault = true,
+                        taskCount = 0,
+                        // tint =primaryColor
+                    ),
+                    Category(
+                        id = 4,
+                        title = "Gym",
+                        // arName = "رياضة",
+                        iconRes = "",
+                        isDefault = false,
+                        taskCount = 0,
+                        //   tint = primaryColor
+                    ),
+                    Category(
+                        id = 5,
+                        title = "Entertainment",
+                        // arName = "ترفيه",
+                        iconRes = "",
+                        isDefault = false,
+                        taskCount = 0,
+                        //tint = primaryColor
+
+                    ),
+                    Category(
+                        id = 6,
+                        title = "Cooking",
+                        // arName = "طبخ",
+                        iconRes = "",
+                        isDefault = false,
+                        taskCount = 0,
+                        // tint = primaryColor
+                    ),
+                    Category(
+                        id = 7,
+                        title = "Family & Friends",
+                        // arName = "العائلة والأصدقاء",
+                        iconRes = "",
+                        isDefault = false,
+                        taskCount = 0,
+                        // tint = primaryColor
+                    ),
+                    Category(
+                        id = 8,
+                        title = "Traveling",
+                        // arName = "سفر",
+                        iconRes = "",
+                        isDefault = false,
+                        taskCount = 0,
+                        // tint = primaryColor
+                    ),
+                    Category(
+                        id = 9,
+                        title = "Agriculture",
+                        //arName = "زراعة",
+                        iconRes = "",
+                        isDefault = false,
+                        taskCount = 0,
+                        //   tint = primaryColor
+                    )
+                ),
                 selectedCategory = selectedCategory,
                 onCategorySelected = { selectedCategory = it }
             )
         }
-    }
-}
-
-@Composable
-private fun rememberSampleDomainCategories(): List<Category> {
-    val primaryColor = TudeeTheme.colors.primary.value
-    val secondaryColor = TudeeTheme.colors.secondary.value
-
-    return remember {
-        listOf(
-            Category(
-                id = 1,
-                title = "Education",
-                // arName = "التعليم",
-                iconRes = "",
-                isDefault = true,
-                taskCount = 0,
-               // tint = primaryColor
-            ),
-            Category(
-                id = 2,
-                title = "Shopping",
-               // arName = "التسوق",
-                iconRes = "",
-                isDefault = true,
-                taskCount = 0,
-               // tint = secondaryColor
-            ),
-            Category(
-                id = 3,
-                title = "Medical",
-               // arName = "طبي",
-                iconRes = "",
-                isDefault = true,
-                taskCount = 0,
-               // tint =primaryColor
-            ),
-            Category(
-                id = 4,
-                title = "Gym",
-               // arName = "رياضة",
-                iconRes = "",
-                isDefault = false,
-                taskCount = 0,
-             //   tint = primaryColor
-            ),
-            Category(
-                id = 5,
-                title = "Entertainment",
-               // arName = "ترفيه",
-                iconRes = "",
-                isDefault = false,
-                taskCount = 0,
-                //tint = primaryColor
-
-            ),
-            Category(
-                id = 6,
-                title = "Cooking",
-               // arName = "طبخ",
-                iconRes = "",
-                isDefault = false,
-                taskCount = 0,
-               // tint = primaryColor
-            ),
-            Category(
-                id = 7,
-                title = "Family & Friends",
-               // arName = "العائلة والأصدقاء",
-                iconRes = "",
-                isDefault = false,
-                taskCount = 0,
-               // tint = primaryColor
-            ),
-            Category(
-                id = 8,
-                title = "Traveling",
-               // arName = "سفر",
-                iconRes = "",
-                isDefault = false,
-                taskCount = 0,
-               // tint = primaryColor
-            ),
-            Category(
-                id = 9,
-                title = "Agriculture",
-                //arName = "زراعة",
-                iconRes = "",
-                isDefault = false,
-                taskCount = 0,
-             //   tint = primaryColor
-            )
-        )
     }
 }
