+++ app/src/main/java/com/london/tudee/presentation/screens/task/add_edit_task_bottom_sheet/AddOrEditTaskUiState.kt
@@ -4,7 +4,6 @@ import androidx.annotation.StringRes
 import com.london.tudee.domain.entities.Category
 import com.london.tudee.domain.entities.Priority
 
-
 data class AddOrEditTaskUiState(
     val taskId: Int? = null,
     val title: String = "",
@@ -18,8 +17,6 @@ data class AddOrEditTaskUiState(
     val showDatePicker: Boolean = false,
     val isLoading: Boolean = false,
     val isFormValid: Boolean = false,
-    @StringRes val successMessage: Int? = null,
-    @StringRes val errorMessage: Int? = null,
     @StringRes val stateMessage: Int? = null,
     val isEditMode: Boolean = false,
 )
