+++ app/src/main/java/com/london/tudee/presentation/screens/task/taskdetails/TaskDetailsBottomSheet.kt
@@ -31,6 +31,7 @@ import com.london.tudee.presentation.components.buttons.TudeeSecondaryButton
 import com.london.tudee.presentation.components.priority.PriorityBadge
 import com.london.tudee.presentation.design_system.theme.ThemePreviews
 import com.london.tudee.presentation.design_system.theme.TudeeTheme
+import com.london.tudee.presentation.screens.task.add_edit_task_bottom_sheet.AddOrEditTaskBottomSheet
 import com.london.tudee.presentation.utils.converterStringToBitmap
 import org.koin.androidx.compose.koinViewModel
 
@@ -56,7 +57,12 @@ fun TaskDetailsBottomSheet(
         onDismiss = onDismiss,
     )
     if (taskDetailsUiState.isEditBottomSheetVisible) {
-        TODO()// edit bottom sheet should be here send task id to it
+        AddOrEditTaskBottomSheet(
+            screenContent = TODO(),
+            taskId = taskId,
+            uiState = TODO(),
+            interactions = TODO()
+        )
     }
 }
